Changed loginattempts to smallint http://community.mybb.com/thread-166156.html

+++ install/resources/mysql_db_tables.php
@@ -1113,7 +1113,7 @@ $tables[] = "CREATE TABLE mybb_users (
   suspendsigtime int unsigned NOT NULL default '0',
   coppauser tinyint(1) NOT NULL default '0',
   classicpostbit tinyint(1) NOT NULL default '0',
-  loginattempts tinyint(2) NOT NULL default '1',
+  loginattempts smallint(2) unsigned NOT NULL default '1',
   usernotes text NOT NULL,
   sourceeditor tinyint(1) NOT NULL default '0',
   UNIQUE KEY username (username),
